StreamlitLLMs/EvaluateQ&AStreamlitLLM/requirements.txt
ipykernel
python == 3.11.4
python-dotenv
streamlit >= 1.37.0
langchain >= 0.2.11
langchain-openai >= 0.1.19
chromadb >= 0.5.5
tiktoken >= 0.7.0
faiss-cpu >= 1.8.0.post1
langchain-community >= 0.2.10
